backend/requirements.txt
# ====================================
# AI Test Automation - Requirements
# Supporta Azure OpenAI
# ====================================

# Core Flask
python-dotenv==1.0.1
flask==3.1.2
flask-cors==5.0.0

# Browser Automation
playwright==1.49.1
playwright-stealth==1.0.6  # Anti-bot detection
setuptools>=65.0.0 # Per playwright-stealth

# LangChain & LangGraph (AI Agent)
langchain==0.3.21
langchain-core==0.3.47
langchain-openai==0.3.9  # Supporta sia OpenAI che Azure
langgraph==0.4.3
langgraph-prebuilt==0.2.2
langgraph-checkpoint==2.0.25

# MCP (Model Context Protocol)
mcp==1.12.3
langchain-mcp-adapters==0.1.7

# Utility
requests==2.32.3
aiofiles==23.2.1
uvicorn==0.34.0

# Per MCP Server HTTP
fastapi==0.115.12
httpx==0.28.1

# Document Processing (Word, HTML)
beautifulsoup4==4.14.3
chardet==7.0.1
python-docx==1.2.0
lxml==6.0.2
